app/requirements.txt
flask
flask-cors
torch
numpy
opencv-python
librosa
mediapipe
transformers
spacy
whisper
soundfile
ffmpeg-python
PyAudio
requests